+++ public/css/signup.css
@@ -1,132 +1,95 @@
 @import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@100;300;400&display=swap');
 
-*{
+* {
     margin: 0;
     padding: 0;
     box-sizing: border-box;
 }
-body{
+
+body {
     font-family: 'Montserrat', sans-serif;
 }
-html{
+
+html {
     scroll-behavior: smooth;
 }
-a{
+
+a {
     text-decoration: none;
 }
-li{
+
+li {
     list-style-type: none;
 }
 
-.main{
+.main {
     width: 60%;
     height: 80vh;
     margin: auto;
     display: flex;
     justify-content: center;
     align-items: center;
-    padding-top: 40px;
+    padding-top: 30px;
 }
 
-.left{
+.right {
     width: 50%;
     height: 100%;
 }
 
-.right{
-    width: 50%;
-    height: 100%;
-}
-
-.left .phone_img{
-    width: 90%;
-    position: relative;
-}
-
-.left .phone_img img{
-    width: 100%;
-}
-
-.left .img_slider{
-    width: 54%;
-    position: absolute;
-    top: 22px;
-    right: 53px;
-    bottom: 0;
-    overflow: hidden;
-    border-radius: 22px;
-}
-
-.left .img_slider .img{
-    width: 100%;
-    overflow: hidden;
-}
-
-.right .form{
-    width: 75%;
+.form {
+    width: 80%; /* Adjust as needed */
+    margin: auto; /* Centers the box horizontally */
     background-color: #ffffff;
     border: 1px solid #dbdbdb;
     border-radius: 1px;
+    padding: 20px;
 }
 
-.right .form .logo{
+.logo {
     width: 165px;
     margin: auto;
-    padding: 35px 0;
 }
 
-.right .form .logo img{
+.logo img {
     width: 100%;
 }
 
-form{
-    width: 80%;
-    margin: auto;
-}
-
-form input{
-    width: 100%;
-    display: block;
-    padding: 10px;
-    margin-bottom: 8px;
-    border-radius: 2px;
-    border: 1px solid #dbdbdb;
-    background-color: #fafafa;
-}
-
-form input::placeholder{
-    font-size: 13px;
-}
-
-form input:focus{
-    outline: none;
-    border: 1px solid #c3c0c0;
-}
-
-form button{
-    background-color: #0095f6;
-    cursor: pointer;
-    width: 100%;
+.logo_text {
     text-align: center;
-    border: none;
-    outline: none;
-    padding: 8px;
-    border-radius: 7px;
-    color: white;
+    margin-bottom: 20px;
     font-weight: 600;
-    margin: 4px 0 15px 0;
+    font-size: 14px;
 }
 
-.form .line{
+.login_by_fb {
+    font-size: 13px;
+    color: #fff; /* Change the text color to white for better contrast */
+    font-weight: bold;
+    display: block;
+    text-align: center;
+    margin-bottom: 10px;
+    background-color: #0095f6; /* Facebook blue background */
+    border-radius: 5px; /* Add border radius */
+    padding: 10px; /* Add padding for better button appearance */
+    text-decoration: none; /* Remove underline from the link */
+}
+
+.fa-square-facebook {
+    margin-right: 5px;
+    font-size: 16px;
+}
+
+.form .line {
     display: block;
     text-align: center;
     position: relative;
     font-size: 13px;
     color: #959393;
-    margin-bottom: 25px;
+    margin: 20px 0;
 }
 
-.form .line::before{
+.line::before{
     content: '';
     display: block;
     width: 30%;
@@ -138,8 +101,7 @@ form button{
     bottom: 0;
     margin: auto;
 }
-
-.form .line::after{
+.line::after {
     content: '';
     display: block;
     width: 30%;
@@ -152,58 +114,96 @@ form button{
     margin: auto;
 }
 
-.login_by_fb{
-    font-size: 14px;
-    color: #385185;
-    font-weight: bold;
-    display: block;
-    text-align: center;
+form {
+    margin-top: 20px;
 }
 
-.fa-square-facebook{
-    margin-right: 5px;
-    font-size: 16px;
+form input {
+    width: 100%;
+    display: block;
+    padding: 10px;
+    margin-bottom: 15px;
+    border-radius: 2px;
+    border: 1px solid #dbdbdb;
+    background-color: #fafafa;
 }
 
-.forgot_password{
-    padding: 20px 0;
-    display: block;
-    text-align: center;
+form input::placeholder {
     font-size: 13px;
-    color: #00376b;
 }
 
-.signup{
-    width: 75%;
+form input:focus {
+    outline: none;
+    border: 1px solid #c3c0c0;
+}
+
+form #signup {
+    background-color: #0095f6;
+    cursor: pointer;
+    width: 100%;
+    text-align: center;
+    border: none;
+    outline: none;
+    padding: 8px;
+    border-radius: 7px;
+    color: white;
+    font-weight: 600;
+    margin-top: 15px;
+    font-size: 13px;
+}
+
+.disclaimer {
+    text-align: center;
+    margin-top: 20px;
+}
+
+.disclaimer p {
+    font-size: 12px;
+    color: #262626;
+    margin-bottom: 5px;
+}
+
+.disclaimer a {
+    color: #0095f6;
+    font-weight: 600;
+}
+
+#sign-up-terms{
+    margin-top:20px;
+}
+
+.login-box{
+    width: 80%;
     background-color: #ffffff;
     border: 1px solid #dbdbdb;
-    padding: 20px 20px;
+    padding: 20px;
+    margin: auto;
     margin-top: 10px;
 }
 
-.signup p{
+.login-box p{
     font-size: 13px;
     color: #262626;
     margin: auto;
     text-align: center;
 }
 
-.signup p a{
+.login-box p a{
     color: #0095f6;
     font-weight: 600;
 }
 
-.right p{
-    width: 75%;
+.get-app{
+    margin: auto;
     text-align: center;
-    font-size: 14px;
-    padding: 10px 0;
+    margin-top: 20px;
 }
 
 .other_socials{
-    width: 70%;
-    padding: 0 30px;
+    width: 80%;
+    padding: 20px;
     display: flex;
+    margin: auto;
     justify-content: space-evenly;
 }
 
@@ -217,19 +217,19 @@ form button{
     height: 100%;
 }
 
-footer{
+footer {
     width: 60%;
     margin: auto;
     text-align: center;
-    padding-top: 120px;
+    padding-top: 400px;
 }
 
-footer ul{
+footer ul {
     display: flex;
     padding-bottom: 10px;
 }
 
-footer ul li a{
+footer ul li a {
     cursor: pointer;
     font-size: 12px;
     padding: 0 8px;
@@ -237,20 +237,20 @@ footer ul li a{
     transition: 0.3s;
 }
 
-footer ul li a:hover{
+footer ul li a:hover {
     text-decoration: underline;
 }
 
-footer .en{
+footer .en {
     font-size: 14px;
     color: #949d9d;
 }
 
-.fa-angle-down{
+.fa-angle-down {
     font-size: 12px;
 }
 
-footer small{
+footer small {
     color: #9f9d9d;
     padding-left: 5px;
 }
