+++ public/html/signup.html
@@ -6,50 +6,39 @@
     <title>Instagram</title>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.2/css/all.min.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/OwlCarousel2/2.3.4/assets/owl.carousel.min.css">
-    <link rel="stylesheet" href="../css/login.css">
+    <link rel="stylesheet" href="../css/signup.css">
 </head>
 <body>
     <div class="main">
-        <div class="left">
-            <div class="phone_img">
-                <img src="../assets/login/home-phones.png">
-                <div class="img_slider owl-carousel">
-                    <div class="img">
-                        <img src="../assets/login/screenshot1.png">
-                    </div>
-                    <div class="img">
-                        <img src="../assets/login/screenshot2.png">
-                    </div>
-                    <div class="img">
-                        <img src="../assets/login/screenshot3.png">
-                    </div>
-                    <div class="img">
-                        <img src="../assets/login/screenshot4.png">
-                    </div>
-                </div>
-            </div>
-        </div>
-
         <div class="right">
             <div class="form">
                 <div class="logo">
                     <img src="../assets/login/login-logo.png">
                 </div>
-                <form action="#">
-                    <input type="text" id="username" placeholder="Phone number, username or email">
-                    <input type="password" id="password" placeholder="Password">
-                    <button type="submit" id="login">Log In</button>
-                </form>
-                <span class="line">OR</span>
+                <div class="logo_text">
+                    <p>Sign up to see photos and videos from your friends.</p>
+                </div>
                 <a class="login_by_fb" href="https://m.facebook.com/login.php?skip_api_login=1&api_key=124024574287414&kid_directed_site=0&app_id=124024574287414&signed_next=1&next=https%3A%2F%2Fm.facebook.com%2Fdialog%2Foauth%3Fclient_id%3D124024574287414%26locale%3Den_US%26redirect_uri%3Dhttps%253A%252F%252Fwww.instagram.com%252Faccounts%252Fsignup%252F%26response_type%3Dcode%252Cgranted_scopes%26scope%3Demail%26state%3D%257B%2522fbLoginKey%2522%253A%252218onj9s1qghsowvuat4gadzl7j1ntsrth1nv5ug0pf0vt51om320s%2522%252C%2522fbLoginReturnURL%2522%253A%2522%252Ffxcal%252Fdisclosure%252F%253Fnext%253D%25252F%2522%257D%26ret%3Dlogin%26fbapp_pres%3D0%26logger_id%3Dfcfb3b2f-ba97-43dc-a06f-fd3782646775%26tp%3Dunspecified&cancel_url=https%3A%2F%2Fwww.instagram.com%2Faccounts%2Fsignup%2F%3Ferror%3Daccess_denied%26error_code%3D200%26error_description%3DPermissions%2Berror%26error_reason%3Duser_denied%26state%3D%257B%2522fbLoginKey%2522%253A%252218onj9s1qghsowvuat4gadzl7j1ntsrth1nv5ug0pf0vt51om320s%2522%252C%2522fbLoginReturnURL%2522%253A%2522%252Ffxcal%252Fdisclosure%252F%253Fnext%253D%25252F%2522%257D%23_%3D_&display=touch&locale=en_GB&pl_dbl=0&refsrc=deprecated&_rdr">
                     <i class="fa-brands fa-square-facebook"></i>Log in with Facebook
                 </a>
-                <a class="forgot_password" href="#">Forgot password?</a>
+                <span class="line">OR</span>
+                <form action="#">
+                    <input type="text" id="email" placeholder="Mobile number or Email">
+                    <input type="text" id="fullname" placeholder="Full Name">
+                    <input type="text" id="username" placeholder="Username">
+                    <input type="password" id="password" placeholder="Password">
+                    <div class="disclaimer">
+                        <p>People who use our service may have uploaded your contact information to Instagram. <a href="#">Learn More</a></p>
+                        <p id="sign-up-terms">By signing up, you agree to our <a href="#">Terms</a>, <a href="#">Privacy Policy</a> and <a href="#">Cookies Policy</a>.</p>
+                    </div>
+                    <button type="submit" id="signup">Sign Up</button>
+                </form>
             </div>
-            <div class="signup">
-                <p><a href="../html/signup.html" id="signup">Create new account</a></p>
+
+            <div class="login-box">
+                <p>Have an account? <a href="../html/login.html" id="login-box-link">Log in</a></p>
             </div>
-            <p>Get the app.</p>
+            <p class="get-app">Get the app.</p>
             <div class="other_socials">
                 <div class="img">
                     <img src="../assets/login/google_play.png" onclick="openPlayStore()">
@@ -83,7 +72,7 @@
     </footer>
 
     <script src="../assets/cdn-js/jquery-3.7.1.min.js"></script>
-    <script type="module" src="../js/login.js"></script>
+    <script type="module" src="../js/signup.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/OwlCarousel2/2.3.4/owl.carousel.min.js"></script>
 </body>
 </html>
