+++ public/js/login.js
@@ -8,3 +8,32 @@ $(document).ready(function(){
         animateOut: 'fadeOut'
     });
 })
+
+import { auth } from './firebaseConfig.js';
+import { signInWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/10.12.0/firebase-auth.js";
+
+document.getElementById("login").addEventListener("click", function(event) {
+  event.preventDefault();
+  const email = document.getElementById("username").value;
+  const password = document.getElementById("password").value;
+  signInWithEmailAndPassword(auth, email, password)
+    .then((userCredential) => {
+      const user = userCredential.user;
+      localStorage.setItem("email", email);
+      alert("Login success");
+      window.location.href="../html/home.html";
+    })
+    .catch((error) => {
+      const errorMessage = error.message;
+      alert(errorMessage);
+    });
+    
+});
+
+window.openPlayStore = function() {
+  window.location.href = "https://play.google.com/store/search?q=instagram&c=apps&hl=en";
+};
+
+window.openMicrosoftStore = function() {
+  window.location.href = "https://apps.microsoft.com/en-us/detail/9nblggh5l9xt";
+};
